// backend/apps/nestar-api/src/libs/enums/property.enum.ts
import { registerEnumType } from '@nestjs/graphql';

export enum PropertyType {
	PHONE = 'PHONE',
	LAPTOP = 'LAPTOP',
	TABLET = 'TABLET',
	MONITOR = 'MONITOR',
	HEADPHONES = 'HEADPHONES',
	CHARGER = 'CHARGER',
	ACCESSORY = 'ACCESSORY',
	OTHER = 'OTHER',
}

registerEnumType(PropertyType, {
	name: 'PropertyType',
});

export enum PropertyStatus {
	ACTIVE = 'ACTIVE',
	SOLD = 'SOLD',
	DELETE = 'DELETE',
}
registerEnumType(PropertyStatus, {
	name: 'PropertyStatus',
});

export enum PropertyLocation {
	SEOUL = 'SEOUL',
	BUSAN = 'BUSAN',
	INCHEON = 'INCHEON',
	DAEGU = 'DAEGU',
	GYEONGJU = 'GYEONGJU',
	GWANGJU = 'GWANGJU',
	CHONJU = 'CHONJU',
	DAEJON = 'DAEJON',
	JEJU = 'JEJU',
}
registerEnumType(PropertyLocation, {
	name: 'PropertyLocation',
});